+++ cpp/examples.cpp
